pyproject.toml
[tool.poetry]
name = "heureka-task-sj"
version = "0.1.0"
description = ""
authors = ["sara-jura <sara.jurankova@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pika = "^1.3.2"
psycopg2-binary = "^2.9.9"

[tool.poetry.dev-dependencies]
pytest = "^7.4.3"
isort = "^5.13.1"
black = "^23.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
